pip-requirements.txt
django>=1.3
django-videos
django-tagging

# for tests
-e git+https://github.com/nose-devs/nose.git#egg=nose
django-nose
coverage
-e git+https://github.com/pydevua/django-testrecorder.git#egg=django-testrecorder